refactor: decompose CapabilityMatchSupervisor (471→187+89+219) into partials
Split by responsibility: delegation entry, capability matching, escalation handling.
