+++ src/Content/Infrastructure/Infrastructure.AI/Agents/CapabilityMatchSupervisor.cs
@@ -27,7 +27,7 @@ namespace Infrastructure.AI.Agents;
 /// pluggable agent selection and <see cref="IDelegationStore"/> for append-only
 /// delegation lifecycle tracking.
 /// </summary>
-public sealed class CapabilityMatchSupervisor : ISupervisor, IDisposable
+public sealed partial class CapabilityMatchSupervisor : ISupervisor, IDisposable
 {
     private const string SupervisorId = nameof(CapabilityMatchSupervisor);
 
@@ -183,289 +183,5 @@ public sealed class CapabilityMatchSupervisor : ISupervisor, IDisposable
         foreach (var cts in _activeDelegations.Values)
             cts.Dispose();
         _activeDelegations.Clear();
-    }
-
-    private async Task<DelegationResult?> HandleAutonomyEscalationAsync(
-        string taskDescription,
-        IReadOnlyList<string> requiredCapabilities,
-        AutonomyLevel minimumTier,
-        int currentDelegationDepth,
-        IReadOnlyList<string>? toolOverrides,
-        Domain.Common.Config.AI.Governance.EscalationConfig escalationConfig,
-        CancellationToken ct)
-    {
-        var escalationRequest = new EscalationRequest
-        {
-            EscalationId = Guid.NewGuid(),
-            AgentId = SupervisorId,
-            ToolName = $"delegate:{string.Join(",", requiredCapabilities)}",
-            Arguments = new Dictionary<string, string>
-            {
-                ["taskDescription"] = taskDescription,
-                ["minimumTier"] = minimumTier.ToString()
-            },
-            Description = $"Delegation blocked by autonomy tier ({minimumTier}): {taskDescription}",
-            RiskLevel = RiskLevel.Medium,
-            Priority = EscalationPriority.Blocking,
-            ApprovalStrategy = Enum.TryParse<ApprovalStrategyType>(
-                escalationConfig.DefaultApprovalStrategy, true, out var strategy)
-                ? strategy : ApprovalStrategyType.AnyOf,
-            Approvers = [],
-            QuorumThreshold = 1,
-            TimeoutSeconds = escalationConfig.DefaultTimeoutSeconds,
-            TimeoutAction = Enum.TryParse<EscalationTimeoutAction>(
-                escalationConfig.DefaultTimeoutAction, true, out var timeoutAction)
-                ? timeoutAction : EscalationTimeoutAction.DenyAndEscalate,
-            RequestedAt = DateTimeOffset.UtcNow
-        };
-
-        if (_escalationService is not { } escalation)
-            return null;
-
-        _logger.LogInformation(
-            "Autonomy tier violation — escalating delegation for {TaskDescription} (minimumTier: {MinimumTier})",
-            taskDescription, minimumTier);
-
-        try
-        {
-            var outcome = await escalation.RequestEscalationAsync(escalationRequest, ct);
-
-            if (!outcome.IsApproved)
-            {
-                _logger.LogWarning("Escalation {EscalationId} denied for delegation: {TaskDescription}",
-                    outcome.EscalationId, taskDescription);
-                return null;
-            }
-
-            _logger.LogInformation("Escalation {EscalationId} approved — retrying delegation with Restricted tier",
-                outcome.EscalationId);
-
-            return await DelegateAsync(
-                taskDescription, requiredCapabilities, AutonomyLevel.Restricted,
-                currentDelegationDepth, toolOverrides, ct);
-        }
-        catch (Exception ex) when (ex is not OperationCanceledException)
-        {
-            _logger.LogError(ex,
-                "Escalation service failed for delegation {TaskDescription} — denying (fail-closed)",
-                taskDescription);
-            return null;
-        }
-    }
-
-    private async Task<SupervisorDecisionContext> BuildDecisionContextAsync(
-        string taskDescription,
-        IReadOnlyList<string> requiredCapabilities,
-        AutonomyLevel minimumTier,
-        int currentDepth,
-        int maxDepth,
-        CancellationToken ct)
-    {
-        var profiles = _profileRegistry.GetAllProfiles();
-        var candidates = new List<AgentCandidate>(profiles.Count);
-
-        foreach (var (type, definition) in profiles)
-        {
-            var tools = _toolResolver.ResolveToolsForSubagent(definition, Array.Empty<AITool>());
-            var toolNames = tools.Select(t => t.Name).ToList();
-            var autonomy = _tierResolver.Resolve(definition);
-
-            candidates.Add(new AgentCandidate
-            {
-                AgentId = type.ToString(),
-                AgentType = type,
-                AutonomyLevel = autonomy,
-                AvailableTools = toolNames
-            });
-        }
-
-        Domain.AI.Routing.Models.TaskComplexityAssessment? complexityAssessment = null;
-
-        if (_modelRouter is not null)
-        {
-            try
-            {
-                complexityAssessment = await _modelRouter.AssessTaskComplexityAsync(
-                    taskDescription, requiredCapabilities, ct);
-            }
-            catch (Exception ex) when (ex is not OperationCanceledException)
-            {
-                _logger.LogWarning(ex,
-                    "Model router complexity assessment failed — proceeding without complexity hint");
-            }
-        }
-
-        return new SupervisorDecisionContext
-        {
-            TaskDescription = taskDescription,
-            RequiredCapabilities = requiredCapabilities,
-            MinimumAutonomyLevel = minimumTier,
-            AvailableAgents = candidates,
-            CurrentDelegationDepth = currentDepth,
-            MaxDelegationDepth = maxDepth,
-            ComplexityAssessment = complexityAssessment
-        };
-    }
-
-    private static DelegationRecord BuildPendingRecord(
-        Guid delegationId,
-        AgentSelection selection,
-        string taskDescription,
-        IReadOnlyList<string> requiredCapabilities,
-        IReadOnlyList<string>? toolOverrides,
-        int currentDepth)
-    {
-        return new DelegationRecord
-        {
-            DelegationId = delegationId,
-            SupervisorId = SupervisorId,
-            DelegateAgentId = selection.SelectedAgent.AgentId,
-            DelegateAgentType = selection.SelectedAgent.AgentType,
-            TaskDescription = taskDescription,
-            RequiredCapabilities = requiredCapabilities,
-            ToolOverrides = toolOverrides,
-            AutonomyLevel = selection.SelectedAgent.AutonomyLevel,
-            State = DelegationState.Pending,
-            DelegationDepth = currentDepth,
-            StartedAt = DateTimeOffset.UtcNow
-        };
-    }
-
-    private async Task<DelegationResult> ExecuteAndTrack(
-        Guid delegationId,
-        DelegationRecord pendingRecord,
-        AgentSelection selection,
-        IReadOnlyList<string>? toolOverrides,
-        int currentDepth,
-        int timeoutSeconds,
-        Stopwatch stopwatch,
-        CancellationToken ct)
-    {
-        var acquired = await _concurrencySemaphore.WaitAsync(
-            TimeSpan.FromSeconds(timeoutSeconds), ct);
-
-        if (!acquired)
-        {
-            await RecordFailure(pendingRecord, "Concurrency semaphore acquisition timed out.", ct);
-            return DelegationResult.Fail("Concurrency semaphore acquisition timed out.");
-        }
-
-        using var cts = CancellationTokenSource.CreateLinkedTokenSource(ct);
-        cts.CancelAfter(TimeSpan.FromSeconds(timeoutSeconds));
-        _activeDelegations[delegationId] = cts;
-
-        try
-        {
-            return await ExecuteAgent(pendingRecord, selection, toolOverrides, currentDepth, stopwatch, cts.Token);
-        }
-        catch (OperationCanceledException)
-        {
-            var reason = ct.IsCancellationRequested ? "Delegation cancelled." : "Delegation timed out.";
-            await RecordCancellation(pendingRecord, reason, ct);
-            return DelegationResult.Fail(reason);
-        }
-        catch (Exception ex)
-        {
-            _logger.LogError(ex, "Delegation {DelegationId} to {AgentId} failed",
-                delegationId, selection.SelectedAgent.AgentId);
-            await RecordFailure(pendingRecord, ex.Message, ct);
-            return DelegationResult.Fail(ex.Message);
-        }
-        finally
-        {
-            _activeDelegations.TryRemove(delegationId, out _);
-            _concurrencySemaphore.Release();
-        }
-    }
-
-    private async Task<DelegationResult> ExecuteAgent(
-        DelegationRecord pendingRecord,
-        AgentSelection selection,
-        IReadOnlyList<string>? toolOverrides,
-        int currentDepth,
-        Stopwatch stopwatch,
-        CancellationToken ct)
-    {
-        var definition = _profileRegistry.GetProfile(selection.SelectedAgent.AgentType);
-        var agentContext = _contextFactory.CreateFromDelegation(definition, toolOverrides, currentDepth + 1, pendingRecord.DelegationId);
-        var agent = await _agentFactory.CreateAgentAsync(agentContext, ct);
-
-        stopwatch.Stop();
-
-        await RecordCompletion(pendingRecord, ct);
-
-        var durationMs = stopwatch.ElapsedMilliseconds;
-
-        SupervisorMetrics.DelegationsTotal.Add(1,
-            new(SupervisorConventions.SupervisorId, SupervisorId),
-            new(SupervisorConventions.DelegateAgentId, selection.SelectedAgent.AgentId),
-            new(SupervisorConventions.Outcome, "completed"));
-
-        SupervisorMetrics.DelegationDuration.Record(durationMs,
-            new(SupervisorConventions.SupervisorId, SupervisorId),
-            new(SupervisorConventions.DelegateAgentId, selection.SelectedAgent.AgentId));
-
-        _auditService.Log(
-            SupervisorId,
-            $"completed:{selection.SelectedAgent.AgentId}",
-            $"delegation {pendingRecord.DelegationId} completed in {durationMs}ms");
-
-        _logger.LogInformation(
-            "Delegation {DelegationId} to {AgentId} completed in {DurationMs}ms",
-            pendingRecord.DelegationId, selection.SelectedAgent.AgentId, durationMs);
-
-        return DelegationResult.Success(
-            $"Agent {selection.SelectedAgent.AgentId} created for delegation {pendingRecord.DelegationId}",
-            0,
-            durationMs);
-    }
-
-    private async Task RecordCompletion(DelegationRecord pendingRecord, CancellationToken ct)
-    {
-        var record = pendingRecord with
-        {
-            State = DelegationState.Completed,
-            CompletedAt = DateTimeOffset.UtcNow
-        };
-
-        await _delegationStore.AppendAsync(record, ct);
-    }
-
-    private async Task RecordFailure(DelegationRecord pendingRecord, string reason, CancellationToken ct)
-    {
-        _auditService.Log(SupervisorId, $"failed:{pendingRecord.DelegateAgentId}", reason);
-
-        SupervisorMetrics.DelegationsTotal.Add(1,
-            new(SupervisorConventions.SupervisorId, SupervisorId),
-            new(SupervisorConventions.DelegateAgentId, pendingRecord.DelegateAgentId),
-            new(SupervisorConventions.Outcome, "failed"));
-
-        var record = pendingRecord with
-        {
-            State = DelegationState.Failed,
-            CompletedAt = DateTimeOffset.UtcNow,
-            FailureReason = reason
-        };
-
-        await _delegationStore.AppendAsync(record, ct);
-    }
-
-    private async Task RecordCancellation(DelegationRecord pendingRecord, string reason, CancellationToken ct)
-    {
-        _auditService.Log(SupervisorId, $"cancelled:{pendingRecord.DelegateAgentId}", reason);
-
-        SupervisorMetrics.DelegationsTotal.Add(1,
-            new(SupervisorConventions.SupervisorId, SupervisorId),
-            new(SupervisorConventions.DelegateAgentId, pendingRecord.DelegateAgentId),
-            new(SupervisorConventions.Outcome, "cancelled"));
-
-        var record = pendingRecord with
-        {
-            State = DelegationState.Cancelled,
-            CompletedAt = DateTimeOffset.UtcNow,
-            FailureReason = reason
-        };
-
-        await _delegationStore.AppendAsync(record, ct);
     }
 }
